bring `build()` back
